\name{bnpspatial-package}
\alias{bnpspatial-package}
\alias{bnpspatial}
\docType{package}
\title{
\packageTitle{bnpspatial}
}
\description{
\packageDescription{bnpspatial}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bnpspatial}
\packageIndices{bnpspatial}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bnpspatial}

Maintainer: \packageMaintainer{bnpspatial}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
